compras y herencia

+++ src/es/avalon/objetos/ejemplo17/Matematicas.java
@@ -2,22 +2,22 @@ package es.avalon.objetos.ejemplo17;
 
 public class Matematicas {
 
-    public  static double sumar (int n1, int n2) {
+    public  static double sumar (double n1, double n2) {
 
         return n1+n2;
     }
 
-    public  static double restar (int n1, int n2) {
+    public  static double restar (double n1, double n2) {
 
         return n1-n2;
     }
 
-    public static double producto (int n1, int n2) {
+    public static double producto (double  n1, double n2) {
 
         return n1*n2;
     }
 
-    public static  double division (int n1, int n2) {
+    public static  double division (double n1, double  n2) {
 
         return n1/n2;
     }
